dags-test-sync/requirements.txt
apache-airflow
apache-airflow-providers-cncf-kubernetes
pandas
numpy